add select option(wget, request) (7z, zipfile ilb)

+++ service/remote/request.py
@@ -5,6 +5,19 @@ import subprocess
 
 from config.app_config import D_SUCCESS, D_ERROR
 
+# subprocess
+def request_subprocess(logger, download_cmd, retry_max, retry_sleep):
+    rtn = D_ERROR
+    for _ in range(retry_max):
+        rtn = subprocess_run(logger, download_cmd)
+        if rtn == 0:  # if download success, subprocess 0 returned.
+            rtn = D_SUCCESS
+            break
+        else:
+            logger.warn(f"Executed retry of file collection from ESP")
+            time.sleep(retry_sleep)
+
+    return rtn
 
 def esp_download(logger, url, fname, timeout, twofactor, retry_max, retry_sleep):
     rtn = D_ERROR
@@ -21,25 +34,25 @@ def esp_download(logger, url, fname, timeout, twofactor, retry_max, retry_sleep)
     return rtn
 
 
-def esp_upload(logger, url, heaer, file, fname, timeout, twofactor, retry_max, retry_sleep):
+def esp_upload(logger, url, heaer, file, timeout, twofactor, retry_max, retry_sleep):
     rtn = D_ERROR
 
-    for _ in range(retry_max + 1):
-        rtn = _request_multipart(url, heaer, file, fname, timeout, twofactor)
+    for _ in range(retry_max):
+        rtn = _request_multipart(url, heaer, file, timeout, twofactor)
         if rtn == D_SUCCESS:
             break
         else:
-            logger.warn(f"Upload Failed. {fname}.")
+            logger.warn(f"Upload Failed. {file}.")
             time.sleep(retry_sleep)
 
 
-def _request_multipart(url, header, file, fname, timeout, twofactor):
+def _request_multipart(url, header, file, timeout, twofactor):
     try:
         if len(twofactor):
             verify = twofactor["cacert"]
             cert = (twofactor["cert"], twofactor["key"])
             connect_read_timeout = (timeout, None)
-            res = requests.post(url, header=header, files=None, timeout=connect_read_timeout, stream=True,
+            res = requests.post(url, headers=header, files=file, timeout=connect_read_timeout, stream=True,
                                 verify=verify, cert=cert)
         else:
             connect_read_timeout = (timeout, None)
@@ -50,14 +63,6 @@ def _request_multipart(url, header, file, fname, timeout, twofactor):
 
     if response.status_code != 200:
         return D_ERROR
-
-    # shlee todo 로그 쌓아야함!!!!!! 다른 업로드들이랑 중복되는지 확인 후 수정
-    # UploadBatch/Upload_Tool.bat - 180
-    # if exist % LOGOPTION % (                                      # LOGOPTION 존재하면
-    #         copy % LOGOPTION % +result.tmp % LOGOPTION %          # LOGOPTION + result.tmp 내용을 LOGOPTION 으로 덮기
-    # ) else (
-    # copy result.tmp % LOGOPTION %                                 # 아님 result.tmp 를 LOGOPTION으로 덮기
-    # )
     return D_SUCCESS
 
 
@@ -81,6 +86,9 @@ def _request_inner(url, fname, timeout, twofactor):
         return D_ERROR
 
     try:
+        if fname is None:
+            return D_SUCCESS
+
         if os.path.exists(fname):
             os.remove(fname)
 
@@ -98,20 +106,6 @@ def _request_inner(url, fname, timeout, twofactor):
 
     return D_SUCCESS
 
-
-# subprocess
-def esp_download_subprocess(logger, download_cmd, retry_max, retry_sleep):
-    rtn = D_ERROR
-    for _ in range(retry_max):
-        rtn = subprocess_run(logger, download_cmd)
-        if rtn == 0:  # if download success, subprocess 0 returned.
-            rtn = D_SUCCESS
-            break
-        else:
-            logger.warn(f"Executed retry of file collection from ESP")
-            time.sleep(retry_sleep)
-
-    return rtn
 
 
 def subprocess_run(logger, cmd):
